cart works, currency symbol integrated

+++ features/step_definitions/product_steps.rb
@@ -1,5 +1,5 @@
 Given /^A product exists with name "([^"]*)" and price "([^"]*)"$/ do |name, price|
-  Product.create!(:name => name, :price => price.to_f)
+  Product.create!(:name => name, :price => price.to_i)
 end
 
 When /^I go to the products page$/ do
